third_party/blink/web_tests/platform/generic/http/tests/devtools/text-source-map-expected.txt

Verify TextSourceMap implementation


Running: testSimpleMap
example.js === example.js
0 === 0
9 === 9
example.js === example.js
0 === 0
13 === 13
example.js === example.js
0 === 0
25 === 25
example.js === example.js
2 === 2
4 === 4
example.js === example.js
2 === 2
11 === 11
example.js === example.js
2 === 2
24 === 24
undefined === undefined
undefined === undefined
undefined === undefined
0 === 0
0 === 0
0 === 0
17 === 17
0 === 0
18 === 18
source line 4 has no mappings.
0 === 0
29 === 29

Running: testNoMappingEntry
example.js === example.js
0 === 0
0 === 0
example.js === example.js
0 === 0
2 === 2

Running: testEmptyLine
example.js === example.js
0 === 0
0 === 0
3 === 3
1 === 1

Running: testNonSortedMappings
example.js === example.js
0 === 0
3 === 3
example.js === example.js
0 === 0
2 === 2
example.js === example.js
0 === 0
1 === 1
example.js === example.js
0 === 0
0 === 0

Running: testSections
source1.js === source1.js
0 === 0
0 === 0
source1.js === source1.js
2 === 2
1 === 1
source2.js === source2.js
0 === 0
0 === 0
source2.js === source2.js
2 === 2
1 === 1

Running: testSourceRoot
/example.js === /example.js
0 === 0
9 === 9
0 === 0
0 === 0

Running: testNameClash
example.js === example.js? [sm]
0 === 0
9 === 9
source line 0 has no mappings.

Running: testNameIndexes
1:0 > [no name assigned]
3:0 > [no name assigned]
3:4 > 'name1'
3:18 > [no name assigned]
4:0 > [no name assigned]
4:0 > [no name assigned]
4:4 > 'generated31465'
4:27 > [no name assigned]
5:0 > [no name assigned]
5:0 > [no name assigned]
